requirements.txt
flask
gunicorn
tensorflow-cpu==2.6.0
opencv-contrib-python-headless==4.5.3.56
numpy==1.19.5
